feat(prometheus): implement metrics prometheus endpoint instead of the status
