+++ frontend/package.json
@@ -68,6 +68,7 @@
     "echarts-for-react": "^3.0.1",
     "fast-deep-equal": "^3.1.3",
     "hexy": "^0.3.1",
+    "json-2-prom": "1.0.3",
     "lodash": "^4.17.21",
     "mixpanel-browser": "^2.41.0",
     "moment": "^2.29.1",
